+++ vimwiki_cli/__init__.py
@@ -21,4 +21,4 @@
 # OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
 # SUCH DAMAGE.
 
-__version__ = '1.0.2'
+__version__ = '1.1.0'